airflow/requirements.txt
apache-airflow[celery,postgres,crypto]==2.9.1
geopandas
pandas
numpy
folium
pdfplumber
pyyaml
seaborn
matplotlib
